Fix feature flags
